extract readline internals
