+++ src/js/node/readline.ts
@@ -27,7 +27,7 @@
 // ----------------------------------------------------------------------------
 const EventEmitter = require("node:events");
 const { StringDecoder } = require("node:string_decoder");
-const { CSI } = require("internal/readline/utils");
+const { CSI, utils: { getStringWidth, stripVTControlCharacters } } = require("internal/readline/utils");
 const { kClearLine, kClearScreenDown, kClearToLineBeginning, kClearToLineEnd } = CSI;
 
 const {
@@ -40,7 +40,6 @@ const {
   validateUint32,
 } = require("internal/validators");
 
-const internalGetStringWidth = $newZigFunction("string.zig", "String.jsGetStringWidth", 1);
 
 const ObjectGetPrototypeOf = Object.getPrototypeOf;
 const ObjectGetOwnPropertyDescriptors = Object.getOwnPropertyDescriptors;
@@ -74,7 +73,6 @@ const ArrayPrototypeReverse = Array.prototype.reverse;
 const ArrayPrototypeShift = Array.prototype.shift;
 const ArrayPrototypeUnshift = Array.prototype.unshift;
 const RegExpPrototypeExec = RegExp.prototype.exec;
-const RegExpPrototypeSymbolReplace = RegExp.prototype[SymbolReplace];
 const StringFromCharCode = String.fromCharCode;
 const StringPrototypeCharCodeAt = String.prototype.charCodeAt;
 const StringPrototypeCodePointAt = String.prototype.codePointAt;
@@ -131,34 +129,6 @@ var SafeStringIterator = createSafeIterator(StringPrototypeSymbolIterator, Strin
 // ----------------------------------------------------------------------------
 // Section: "Internal" modules
 // ----------------------------------------------------------------------------
-
-/**
- * Returns the number of columns required to display the given string.
- */
-var getStringWidth = function getStringWidth(str, removeControlChars = true) {
-  if (removeControlChars) str = stripVTControlCharacters(str);
-  str = StringPrototypeNormalize.$call(str, "NFC");
-  return internalGetStringWidth(str);
-};
-
-// Regex used for ansi escape code splitting
-// Adopted from https://github.com/chalk/ansi-regex/blob/HEAD/index.js
-// License: MIT, authors: @sindresorhus, Qix-, arjunmehta and LitoMore
-// Matches all ansi escape code sequences in a string
-var ansiPattern =
-  "[\\u001B\\u009B][[\\]()#;?]*" +
-  "(?:(?:(?:(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]+)*" +
-  "|[a-zA-Z\\d]+(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]*)*)?\\u0007)" +
-  "|(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PR-TZcf-ntqry=><~]))";
-var ansi = new RegExp(ansiPattern, "g");
-
-/**
- * Remove all VT control characters. Use to estimate displayed string width.
- */
-function stripVTControlCharacters(str) {
-  validateString(str, "str");
-  return RegExpPrototypeSymbolReplace.$call(ansi, str, "");
-}
 
 // Promisify
 
@@ -2920,14 +2890,6 @@ export default {
     Interface: PromisesInterface,
     createInterface(input, output, completer, terminal) {
       return new PromisesInterface(input, output, completer, terminal);
-    },
-  },
-
-  [SymbolFor("__BUN_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED__")]: {
-    CSI,
-    utils: {
-      getStringWidth,
-      stripVTControlCharacters,
     },
   },
 };
