+++ test/js/node/readline/getStringWidth.test.ts
@@ -1,9 +1,7 @@
-import readline from "node:readline";
-
-var {
+// Flags: --expose-internals
+const {
   utils: { getStringWidth },
-  // @ts-ignore
-} = readline[Symbol.for("__BUN_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED__")];
+} = require("internal/readline/utils");
 
 it("handles invisible ASCII character at any position", () => {
   const visible = "a";
